Move disabling install button to if-else codeblock
This ensures that the disabled "install" button is correctly shown iff the percentage calculation is not ready.

This also enables for future changes to disable the button onClick for other cases (like update, reinstall, etc)
Add checks on when to change to downloading status

InstallStatus.Downloading should only be called when the percentage is between 0 to 100, inclusive, and the file is not completely downloaded yet.

Also, move the disabling of button when Install is tapped/clicked at downloadPackages function.
fix